+++ app/code/Magento/Sales/Model/Order/Pdf/Config/Reader.php
@@ -1,0 +1,22 @@
+<?php
+/**
+ * Loads catalog attributes configuration from multiple XML files by merging them together
+ *
+ * {license_notice}
+ *
+ * @copyright   {copyright}
+ * @license     {license_link}
+ */
+class Magento_Sales_Model_Order_Pdf_Config_Reader extends Magento_Config_Reader_Filesystem
+{
+    /**
+     * List of identifier attributes for merging
+     *
+     * @var array
+     */
+    protected $_idAttributes = array(
+        '/config/renderers/page' => 'type',
+        '/config/renderers/page/renderer' => 'product_type',
+        '/config/totals/total' => 'name',
+    );
+}